Use corrected runpaths w/ setup.py develop too
This ensures that a previous fix which improved the installation of
cc_sym and symengine when symforce was installed as an editable package
using the new pep 660 build_editable hook is also applied when symforce
is installed with `python setup.py develop`.

Most notably, this applies to all editable installations before
setuptools==64.0.0.