Make sym and other genned pkgs namespace packages
Still a WIP. Right now `pkgutil.extendpath` is a little
inadequate, as it only searches the `sys.path` and doesn't
make use of any of the finders on `sys.meta_path` (which
is how `sym` is found during an editable install if you
uninstall the non-editable install of `sym` that also gets
installed, which you might reasonably do if you actually
want `sym` to be an editable install as well).

I'm thinking I might want to try seeing if any of those
finders can find the spec, and if so, to then append the
spec's `submodule_search_locations` to the `__path__`.

Okay. That last suggestion was done.