DRAFT: Add variant shapes of sqrt_info in factors
Often times the square root information matrix is a diagonal or
isotropic matrix. However, in our generated geo factors we always assume
they are full dense matrices, leading to many more operations than
needed.

This commit modify `geo_factors_codegen.py` to create 3 variants for
each geo factor, one which assumes `sqrt_info` is a square matrix (the
existing behavior), one which assumes it is a diagonal matrix, and one
which assumes it is an isotropic matrix.

These variants are all generated with the same name are distinguished
only by their signatures (thus this only works for C++, which is all we
currently generate factors for anyway). The square version takes a square
eigen matrix, the diagonal version a eigen vector whose entries are
those of the diagonal of `sqrt_info`, and a single scalar for the
isotropic version.

The header that the user imports remains the same, except now rather
than containing the implementation directly, it re-imports the headers
for the 3 different versions (which are stores in sub-directories).