Allow map type matrices in Factor
In an example optimization problem, there seemed to be an
approximately 10% overhead to each call to the Factor's linearization
function to copy the data from the fixed size output matrices passed to
the generated residual to the dynamic sized output matrices in the
factor returned by `HessianFixedDense`.

This commit aims to help eliminate the overhead by allowing a Factor to
be constructed from a function taking `Eigen::Map`s for the residual,
jacobian, hessian, and rhs outputs, and then calling the inner function
directly on the data of the linearization (thereby avoiding a copy).

Still need to:
- Testing
- Improve code organization
- Enable generation of functions using the `Eigen::Map` type.

Chose `Eigen::Map` over `Eigen::Ref` because the latter has overhead,
and over templating on `Eigen::Derived` to avoid having to deal with an
extra template argument and to be explicit about the expeted shape.