Set min version of lazy-object-proxy to 1.9.0
This is because lazy-object-proxy does not have built wheels on pypa for
linux on version 1.8.0, and for some reason running it's setup.py can
fail when done locally.

Also, for some reason, a pre-built wheel for the same package is not
available for python3.11 on bionic (even for version 1.9.0). Note,
lazy-object-proxy can be built on pip 22.3.1 (but not on the current
latest, 23.0.0). Suspect this is a bug on pip's end, but if not, then
it's probably a bug on lazy-object-proxy's end. Removing this test
because we don't want to contort our code around other package's
problems.